ocr_requirements.txt
flask
pytesseract
Pillow